chore(api): move code-utils_test and date-utils_tests in shared
