+++ api/tests/shared/unit/infrastructure/utils/code-utils_test.js
@@ -1,5 +1,5 @@
-import { generateNumericalString } from '../../../../src/shared/infrastructure/utils/code-utils.js';
-import { expect } from '../../../test-helper.js';
+import { generateNumericalString } from '../../../../../src/shared/infrastructure/utils/code-utils.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Utils | code-utils', function () {
   describe('#generateNumericalString', function () {
