+++ api/tests/shared/unit/infrastructure/utils/date-utils_test.js
@@ -2,8 +2,8 @@ import {
   anonymizeGeneralizeDate,
   convertDateValue,
   isValidDate,
-} from '../../../../src/shared/infrastructure/utils/date-utils.js';
-import { expect } from '../../../test-helper.js';
+} from '../../../../../src/shared/infrastructure/utils/date-utils.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Utils | date-utils', function () {
   describe('#isValidDate', function () {
